learning_core/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
cleo==0.6.8
cryptography==35.0.0
Deprecated==1.2.13
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
django-heroku==0.3.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
jwcrypto==1.0
# mysqlclient==2.0.3
oauthlib==3.1.1
paramiko==2.8.0
pastel==0.1.1
Pillow==8.3.2
platformdirs==2.2.0
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.1.0
pylev==1.4.0
PyNaCl==1.4.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.1